requirements.txt
camel-ai[all]>=0.2.0
groq
python-dotenv
numpy
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.0.0